Implemented longValue() and fixed tests

+++ csce331-testing-rational/src/test/java/numbers/RationalTest.java
@@ -616,7 +616,7 @@ public class RationalTest
         assertThat("Integer conversion: 0/25213 = 0/1 -> 0", value5.intValue(), is(0));
         assertThat("Integer conversion: 24351/546 = 44.5989... -> 44", value6.intValue(), is(44));
     }
-    
+
     public void testLongValue() {
         Rational value1 = new Rational(2583, 1);
         Rational value2 = new Rational(1, 2);
@@ -624,13 +624,13 @@ public class RationalTest
         Rational value4 = new Rational(-24352, 2);
         Rational value5 = new Rational(0, 25213);
         Rational value6 = new Rational(24351, 546);
-    
-        assertThat("Long conversion: 2583/1 -> 2583", value1.intValue(), is(2583L));
-        assertThat("Long conversion: 1/2 -> 0", value2.intValue(), is(0L));
-        assertThat("Long conversion: -1/2 -> 0", value3.intValue(), is(0L));
-        assertThat("Long conversion: -24352/2 = -12176/1 -> -12176", value4.intValue(), is(-12176L));
-        assertThat("Long conversion: 0/25213 = 0/1 -> 0", value5.intValue(), is(0L));
-        assertThat("Long conversion: 24351/546 = 44.5989... -> 44", value6.intValue(), is(44L));
+
+        assertThat("Long conversion: 2583/1 -> 2583", value1.longValue(), is(2583L));
+        assertThat("Long conversion: 1/2 -> 0", value2.longValue(), is(0L));
+        assertThat("Long conversion: -1/2 -> 0", value3.longValue(), is(0L));
+        assertThat("Long conversion: -24352/2 = -12176/1 -> -12176", value4.longValue(), is(-12176L));
+        assertThat("Long conversion: 0/25213 = 0/1 -> 0", value5.longValue(), is(0L));
+        assertThat("Long conversion: 24351/546 = 44.5989... -> 44", value6.longValue(), is(44L));
     }
 
     public void testFloatValue() {
